refactor(auth): Relocate into src/routes/auth
