+++ src/index.ts
@@ -1,8 +1,8 @@
-import auth from '@/auth'
 import { notFound } from '@/middleware/not-found'
 import { onError } from '@/middleware/on-error'
 import { securityMiddlewareHandler } from '@/middleware/security'
 import { workerLogger } from '@/middleware/worker-logger'
+import auth from '@/routes/auth'
 import { OpenAPIHono } from '@hono/zod-openapi'
 import { apiReference } from '@scalar/hono-api-reference'
 import type { Context } from 'hono'
