+++ src/routes/auth/change-password.ts
