+++ src/routes/auth/index.ts
