+++ src/routes/auth/logout.ts
