+++ src/routes/auth/register.ts
