+++ src/routes/auth/status.ts
