+++ src/routes/auth/tests/register.test.ts
@@ -1,4 +1,4 @@
-import { registerHandler } from '@/auth/register'
+import { registerHandler } from '@/routes/auth/register'
 import type { Context } from 'hono'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
